// src/utils/types.ts
export interface MovieProps {
  backdrop: string;
  cast: string[];
  classification: string;
  director: string;
  genres: string[];
  id: string;
  imdb_rating: number;
  length: string;
  overview: string;
  poster: string;
  released_on: string;
  slug: string;
  title: string;
}

export interface GenreWithMovies {
  name: string;
  movies: MovieProps[];
}